New version
Adding name in RF impairments structure. It can be convinient to store the output of `init_**` in JSON files but in this case we do not have the type of the structure (and thus the constructor that should be used for reconstruction). Adding a field help to store in a pure text file and reconstruct a new `struct`